# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Presensi Skansaba API",
        "description": "API Documentation for Presensi Skansaba - Student Attendance System",
        "contact": {
            "name": "Presensi Skansaba Support",
            "email": "admin@skansaba.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/profile": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update student profile",
                "operationId": "d682da2fd566e34f597e169ac582ad3f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Merdeka No. 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Student not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/profile/password": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update password",
                "operationId": "141e4afbe7b6c73bcd252f096680adb5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Password berhasil diperbarui."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/profile/photo": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Upload profile photo",
                "operationId": "e8c57835a442e2b1e7766e1ef3af3bc3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Photo uploaded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Foto profil berhasil diperbarui."
                                        },
                                        "data": {
                                            "properties": {
                                                "photo_url": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/storage/profile-photos/abc.jpg"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Delete profile photo",
                "operationId": "6feae43807105c7e878126ea965fb1b1",
                "responses": {
                    "200": {
                        "description": "Photo deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Foto profil berhasil dihapus."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Student": {
                "title": "Student",
                "description": "Student model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "nisn": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "nis": {
                        "type": "string",
                        "example": "12345"
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "male",
                            "female"
                        ],
                        "example": "male"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date",
                        "example": "2005-01-15"
                    },
                    "phone": {
                        "type": "string",
                        "example": "081234567890"
                    },
                    "address": {
                        "type": "string",
                        "example": "Jl. Contoh No. 123"
                    },
                    "entry_year": {
                        "type": "integer",
                        "example": 2023
                    },
                    "classroom": {
                        "$ref": "#/components/schemas/Classroom"
                    },
                    "parent_name": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "parent_phone": {
                        "type": "string",
                        "example": "081234567891"
                    },
                    "has_active_device": {
                        "type": "boolean",
                        "example": true
                    },
                    "device_registered_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T10:00:00Z"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "username": {
                        "type": "string",
                        "example": "johndoe"
                    },
                    "photo": {
                        "type": "string",
                        "example": "https://example.com/storage/profiles/photo.jpg",
                        "nullable": true
                    },
                    "role": {
                        "type": "string",
                        "example": "student"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "Classroom": {
                "title": "Classroom",
                "description": "Classroom model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "X IPA 1"
                    },
                    "grade_level": {
                        "type": "integer",
                        "example": 10
                    },
                    "major": {
                        "type": "string",
                        "enum": [
                            "IPA",
                            "IPS"
                        ],
                        "example": "IPA"
                    },
                    "class_number": {
                        "type": "integer",
                        "example": 1
                    },
                    "academic_year": {
                        "type": "string",
                        "example": "2023/2024"
                    },
                    "students_count": {
                        "type": "integer",
                        "example": 30
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}